\name{print.segmented}
\alias{print.segmented}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print method for the segmented class }
\description{
  Printing the most important feautures of a segmented model.
}
\usage{
print.segmented(x, digits = max(3, getOption("digits") - 3), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class \code{segmented} }
  \item{digits}{ number of digits to be printed }
  \item{\dots}{ arguments passed to other functions }
}
\author{ Vito M.R. Muggeo }

\seealso{ \code{\link{summary.segmented}}, \code{\link{print.summary.segmented}} }

\keyword{ models }
