\name{segmented-package}
\alias{segmented-package}
%\alias{segmented}
\docType{package}
\title{
Segmented relationships in regression models
}
\description{
Estimation of Generalized Linear Models with piecewise linear relationships having a
fixed number of break-points.
}
\details{
\tabular{ll}{
Package: \tab segmented\cr
Type: \tab Package\cr
Version: \tab 0.2-3\cr
Date: \tab 2007-10-26\cr
License: \tab GPL\cr
}
 Package \code{segmented} is aimed to estimate linear and generalized linear models
 having one or more segmented relationships in the linear predictor. Estimates of the slopes and
 of the possibly multiple breakpoints are provided. The package includes testing/estimating
 functions and methods to print, summarize and plot the results.
}
\author{
Vito M.R. Muggeo <vmuggeo@dssm.unipa.it> (thanks to Marco Enea for his partial support)
  }
\references{

Seber, G.A.F. and Wild, C.J. (1989) \emph{Nonlinear Regression}. Wiley, New York.

Muggeo, V.M.R. (2003) Estimating regression models with unknown break-points.
    \emph{Statistics in Medicine} \bold{22}, 3055--3071.

Davies, R.B. (1987) Hypothesis testing when a nuisance parameter is present only under the alternative.
    \emph{Biometrika} \bold{74}, 33--43.
    }

\keyword{ regression }
\keyword{ nonlinear }
